requirements.txt
Django==3.1.1
djangorestframework==3.11.0
django-cors-headers==3.8.0
beautifulsoup4==4.9.0
Pillow==7.1.1
requests==2.23.0